Moved string handler overloads to write and writeline respectively
